+++ examples/transform/transform.cpp
@@ -1,8 +1,10 @@
 #include <fluxgl/fluxgl.h>
+#include <glm/gtc/matrix_transform.hpp>
+
 #include <iostream>
 #include <sstream>
 
-class TextureExample : public fluxgl::App {
+class TransformExample : public fluxgl::App {
     public: using fluxgl::App::App; // Inherit constructors
 
     private:
@@ -10,11 +12,14 @@ class TextureExample : public fluxgl::App {
 
     protected:
         void onInit() override {
-            entity.material.shader = fluxgl::Shader::loadFromFiles("assets/shaders/texture.vert", "assets/shaders/texture.frag");
+            entity.material.shader = fluxgl::Shader::loadFromFiles("assets/shaders/vertex.glsl", "assets/shaders/fragment.glsl");
             entity.material.albedoTextures.push_back(fluxgl::Texture::loadFromFile("assets/textures/container.jpg"));
             entity.material.albedoTextures.push_back(fluxgl::Texture::loadFromFile("assets/textures/awesomeface.png"));
             
             entity.mesh = fluxgl::Mesh::quad();
+
+            entity.transform.scale = {0.5f, 0.5f, 0.5f};
+            entity.transform.rotation = glm::vec3(0.0f, 0.0f, 45.0f);
         }
 
         void onRender() override {
@@ -24,7 +29,7 @@ class TextureExample : public fluxgl::App {
 };
 
 int main() {
-    TextureExample app(800, 600, "Texture Example");
+    TransformExample app(800, 600, "Transform Example");
     app.run();
 
     return 0;